TicketingSolution.Core/Handler/TicketBookingRequestHandler.cs
using TicketingSolution.Core.Model;

namespace TicketingSolution.Core.Handler
{
    public class TicketBookingRequestHandler
    {
        public TicketBookingRequestHandler(DataServices.ITicketBookingService @object)
        {
        }

        public ServiceBookingResult BookService(TicketBookingRequest? bookingRequest)
        {
            if (bookingRequest is null)
            {
                throw new ArgumentNullException(nameof(bookingRequest));
            }

            return new ServiceBookingResult
            {
                Name = bookingRequest.Name,
                Family = bookingRequest.Family,
                Email = bookingRequest.Email
            };
        }
    }
}